Rejig definitions
