--- engine-nextgen/src/main/kotlin/graphql/nadel/enginekt/blueprint/GraphQLBatchHydrationField.kt
@@ -1,10 +1,0 @@
-package graphql.nadel.enginekt.blueprint
-
-data class GraphQLBatchHydrationField(
-    val name: String,
-    val sourceService: String,
-    val pathToSourceField: List<String>,
-    val batchSize: Int,
-    val objectIdentifier: String,
-    val matchByIndex: Boolean,
-)
