+++ engine-nextgen/src/main/kotlin/graphql/nadel/enginekt/blueprint/GraphQLExecutionBlueprint.kt
@@ -1,4 +1,9 @@
 package graphql.nadel.enginekt.blueprint
 
-class GraphQLExecutionBlueprint(
+import graphql.schema.FieldCoordinates
+
+data class GraphQLExecutionBlueprint(
+    val underlyingFields: Map<FieldCoordinates, GraphQLUnderlyingField>,
+    val underlyingTypes: Map<String, GraphQLUnderlyingType>,
+    val artificialFields: Map<FieldCoordinates, GraphQLArtificialField>,
 )
