--- engine-nextgen/src/main/kotlin/graphql/nadel/enginekt/blueprint/GraphQLHydrationField.kt
@@ -1,7 +1,0 @@
-package graphql.nadel.enginekt.blueprint
-
-data class GraphQLHydrationField(
-    val name: String,
-    val sourceService: String,
-    val pathToSourceField: List<String>,
-)
