--- engine-nextgen/src/main/kotlin/graphql/nadel/enginekt/blueprint/GraphQLPullField.kt
@@ -1,6 +1,0 @@
-package graphql.nadel.enginekt.blueprint
-
-class GraphQLPullField(
-    val name: String,
-    val path: List<String>,
-)
